Reorganized some of the types and imports added future event gains value and type
